modified

+++ SORTING/BucketSort.java
@@ -7,13 +7,16 @@ public class BucketSort {
         float [] arr = {0.897f, 0.565f, 0.656f, 0.1234f, 0.665f, 0.32f};
         int nbksBuckets= arr.length;
         HashMap<Integer, List<Float>> BucketSort = new HashMap<>();
+
         for (int i=0;i< arr.length;i++)
             BucketSort.put(i,new LinkedList<>());
 
         for (int i=0; i< arr.length;i++)
-            BucketSort.get((int)(arr[i]*nbksBuckets)).add(arr[i]);
+            BucketSort.get((arr[i]*nbksBuckets)).add(arr[i]);
+
         for (int i=0;i< nbksBuckets;i++)
             Collections.sort(BucketSort.get(i));
+
         System.out.println(BucketSort);
     }
 }
